+++ test/specs/utils_group_concat.spec.js
@@ -21,9 +21,9 @@ describe('utils/group_concat', () => {
 		}, {
 			expression: 'table.b',
 			label: 'collection.b'
-		}], 'collection.');
+		}], 'collection.', 'a');
 
-		expect(gc.expression).to.eql(`CONCAT('[', GROUP_CONCAT(CONCAT_WS('', '[', '"', REPLACE(REPLACE(table.a, '\\\\', '\\\\\\\\'), '"', '\\\\"'), '"', ',', '"', REPLACE(REPLACE(table.b, '\\\\', '\\\\\\\\'), '"', '\\\\"'), '"', ']')), ']')`);
+		expect(gc.expression).to.eql(`CONCAT('[', GROUP_CONCAT(IF(a.id IS NOT NULL, CONCAT_WS('', '[', '"', REPLACE(REPLACE(table.a, '\\\\', '\\\\\\\\'), '"', '\\\\"'), '"', ',', '"', REPLACE(REPLACE(table.b, '\\\\', '\\\\\\\\'), '"', '\\\\"'), '"', ']'), NULL)), ']')`);
 		expect(gc.label).to.eql('collection[a,b]');
 
 	});
@@ -63,9 +63,9 @@ describe('utils/group_concat', () => {
 		const gc = group_concat([{
 			expression: 'table.a',
 			label: 'collection.a'
-		}], 'collection.');
+		}], 'collection.', 'a');
 
-		expect(gc.expression).to.eql(`CONCAT('[', GROUP_CONCAT(CONCAT_WS('', '[', '"', REPLACE(REPLACE(table.a, '\\\\', '\\\\\\\\'), '"', '\\\\"'), '"', ']')), ']')`);
+		expect(gc.expression).to.eql(`CONCAT('[', GROUP_CONCAT(IF(a.id IS NOT NULL, CONCAT_WS('', '[', '"', REPLACE(REPLACE(table.a, '\\\\', '\\\\\\\\'), '"', '\\\\"'), '"', ']'), NULL)), ']')`);
 		expect(gc.label).to.eql('collection[a]');
 
 	});